requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
codecov==2.1.12
colorama==0.4.4
coverage==6.3
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
django-sslserver==0.22
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
jmespath==0.10.0
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
tzdata==2021.5
urllib3==1.26.8
wrapt==1.14.0
zipp==3.7.0
